requirements.txt
pandas
numpy
requests
scikit-learn
joblib
PyYAML
python-dotenv
lightgbm
urllib3<2
